Create Orbits Simulation
The Jupyter Notebook File was too big (101 MB) so I couldn't upload it. I hence have copied all the code into a new file. This file does lack the non-code text which encapsulates the code to make it seem presentable, as seen in the other files under the Lab projects.